refactor: 채팅방, 친구 목록 리렌더링 최적화
