+++ src/components/@layout/SideBar/SideBarFriendItem/SideBarFriendItem.styled.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { lineClamp } from '../SideBarChatList/SideBarChatList.styled';
+import { lineClamp } from '../SideBarChatRoomItem/SideBarChatRoomItem.styled';
 
 export const Wrapper = styled.li`
   display: flex;
